project01/requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
streamlit